+++ src/components/notification/notification.ts
@@ -98,11 +98,10 @@ export class Notification implements OnDestroy, OnInit {
     constructor (
         private notificationService: NotificationService,
         private gtmService: GTMService,
-        @Inject(APP_CONFIG) private appConfig: IAppConfig
-    ) {
-        const app = this.appConfig.gtmAppName || this.appConfig.appBasePath;
-        this.defaultErrorAction = app ? 'Action not provided from app ' + app  : 'Error from NG2UEUtils Notification component';
-    }
+        @Inject(APP_CONFIG) private appConfig: IAppConfig ) {
+            const app = this.appConfig.gtmAppName || this.appConfig.appBasePath;
+            this.defaultErrorAction = app ? 'Action not provided from app ' + app  : 'Error from NG2UEUtils Notification component';
+        }
 
     public show (message: string, type?: string, gtmAction?: string) {
         this.notification = message;
@@ -162,7 +161,7 @@ export class Notification implements OnDestroy, OnInit {
     }
 
     private logUserVisibleError (message: string, gtmAction?: string) {
-        this.gtmService.registerEventTrack(gtmAction || this.defaultErrorAction,
-            { category: 'Edge_userVisibleError', label: message });
+        this.gtmService.registerEventTrack(
+            { action : gtmAction || this.defaultErrorAction, category: 'Edge_userVisibleError', label: message });
     }
 }