+++ src/services/client/client.spec.ts
@@ -1,20 +1,35 @@
-import { TestBed }                      from '@angular/core/testing';
+import { TestBed }    from '@angular/core/testing';
 import {
     XHRBackend,
     Response,
     ResponseOptions,
     RequestMethod,
     Headers,
-    HttpModule
-}                                       from '@angular/http';
-import { MockBackend, MockConnection }  from '@angular/http/testing';
+    HttpModule }      from '@angular/http';
+import {
+    MockBackend,
+    MockConnection }  from '@angular/http/testing';
 
-import { Client }                       from './client';
-import { WindowRef, WindowMock }        from '../window-ref';
+import { Client }     from './client';
+import {
+    APP_CONFIG,
+    IAppConfig }      from '../context/app-config';
+import {GTMService, IGAEvent} from '../context/gtm';
+import {
+    WindowRef,
+    WindowMock }      from '../window-ref';
 
 declare const beforeEach, expect, it, describe;
 
 const someUrl = '/organziations/abc/proxies';
+const apiBasePath = 'apiproducts';
+const appBasePath = 'products';
+const appName = 'ProductsSPA';
+const appConfig: IAppConfig = {
+    apiBasePath: apiBasePath,
+    appBasePath: appBasePath,
+    gtmAppName: appName
+};
 
 const emptyResponse: Response = new Response(
     new ResponseOptions({
@@ -26,11 +41,6 @@ const emptyResponse: Response = new Response(
         url: someUrl
     })
 );
-
-const newResponse = (): Response =>
-    new Response(
-        new ResponseOptions()
-    );
 
 const newForbidden = (): Response =>
     new Response(
@@ -47,19 +57,23 @@ const newForbidden = (): Response =>
 describe('Client', () => {
     let client: Client;
     let backend: MockBackend;
+    let window: WindowMock;
 
     beforeEach(() => {
         TestBed.configureTestingModule({
             imports: [ HttpModule ],
             providers: [
                 Client,
+                GTMService,
                 { provide: XHRBackend, useClass: MockBackend },
+                { provide: APP_CONFIG, useValue: appConfig },
                 { provide: WindowRef, useClass: WindowMock },
             ]
         });
 
         client = TestBed.get(Client);
         backend = TestBed.get(XHRBackend);
+        window = TestBed.get(WindowRef);
     });
 
     it('Injection with Angular', () => {
@@ -73,10 +87,18 @@ describe('Client', () => {
             next => {
                 expect(connection.request.url).toBe(someUrl);
                 expect(connection.request.method).toBe(RequestMethod.Get);
+                expect(window.dataLayer.length).toBe(1);
+                const theEvent: IGAEvent = window.dataLayer[0];
+                expect(theEvent.event).toBe('timing');
+                expect(theEvent.action).toBe(200);
+                expect(theEvent.target).toBe('Edge_APICall');
+                expect(theEvent['target-properties']).toBe(someUrl);
+                expect(theEvent.value).toBeDefined();
+                expect(theEvent.value).toBeGreaterThan(0);
                 done();
             }
         );
-        connection.mockRespond(newResponse());
+        connection.mockRespond(emptyResponse);
     });
     
     it('Empty string response bodies do not break the client', (done) => {
@@ -103,7 +125,7 @@ describe('Client', () => {
                 done();
             }
         );
-        connection.mockRespond(newResponse());
+        connection.mockRespond(emptyResponse);
     });
 
     it('POST translates URL correctly', (done) => {
@@ -116,7 +138,7 @@ describe('Client', () => {
                 done();
             }
         );
-        connection.mockRespond(newResponse());
+        connection.mockRespond(emptyResponse);
     });
 
     it('DELETE translates URL correctly', (done) => {
@@ -129,7 +151,7 @@ describe('Client', () => {
                 done();
             }
         );
-        connection.mockRespond(newResponse());
+        connection.mockRespond(emptyResponse);
     });
 
     it('GET with empty error', (done) => {